增加首頁-飲料卡片+店家詳細頁登入前收藏按鈕判斷，以及整理各頁面console.log

+++ assets/js/index-drinks.js
@@ -27,7 +27,6 @@ document.addEventListener("DOMContentLoaded", () => {
       // 這裡可以調用其他函數，如 getTopSixDrinks、drinkTagPush 和 drinkRenderData
       getTopSixDrinks();
       drinkTagPush(); //組合Tag陣列
-      // checkUserShopCollection();
       drinkRenderData(); //載入預設飲料卡片
       isCollect(); //收藏愛心CSS
     })
@@ -48,7 +47,7 @@ document.addEventListener("DOMContentLoaded", () => {
     const drinkTag2 = []; //去掉逗點
     for (let i = 0; i < drinkTag.length; i++) {
       drinkTag2.push(drinkTag[i].split(","));
-    }
+    };
 
     let tagStr = "";
     let partialTags = []; //保存部分標籤
@@ -81,15 +80,11 @@ document.addEventListener("DOMContentLoaded", () => {
       } else {
         // 如果 drinkId 不存在，初始化為 1
         drinkIdCount[drinkId] = 1;
-      }
-    }
-
-    // 驗證
-    // console.log(drinkIdCount);
+      };
+    };
 
     // 用 Object.entries 將物件轉為陣列
     const drinkIdCountArr = Object.entries(drinkIdCount);
-    // console.log(drinkIdCountArray);
 
     // 使用 sort 由大到小排序
     drinkIdCountArr.sort(function (a, b) {
@@ -99,66 +94,55 @@ document.addEventListener("DOMContentLoaded", () => {
     // 取得排序後的前6個元素
     let topSixDrinkIds = drinkIdCountArr.slice(0, 6);
 
-    // 驗證
-    // console.log(topSixDrinkIds);
-
     // 使用 map 方法取出每個陣列的第一個值（drinkId）
     let topSixIds = topSixDrinkIds.map(function (item) {
       return item[0];
     });
-
-    // 驗證
-    // console.log(topSixIds);
 
     // 找出對應的飲料數據
     drinkData.forEach(function (item) {
       topSixIds.forEach(function (i) {
         if (i == item.id) {
           topSixDrinks.push(item);
-        }
+        };
       });
     });
-    // 驗證
-    // console.log(topSixDrinks);
-  }
+  };
 
   //載入預設飲料卡片函式------------------------------------------------
   const drinkRender = () => {
     let str = "";
     topSixDrinks.forEach((item) => {
-      // console.log(item.id);
-      // console.log(userDrinkCollectionList);
-      // console.log(userDrinkCollectionList.includes(item.id));
       const btn_tag = userDrinkCollectionList.includes(+item.id)
         ? `<button type="button" class="collect-btn collect-drink-btn border-0 text-primary fa-heart fs-24 fa-solid" data-drink-id="${item.id}" value="collected" aria-hidden="true"></button>`
         : `<button type="button" class="collect-btn collect-drink-btn border-0 text-primary fa-regular fa-heart fs-24" data-drink-id="${item.id}" value="uncollect" aria-hidden="true"></button>`;
 
       str += `
-    <li class="drinks-card px-16 py-24 px-md-24">
-    ${btn_tag}
-    <img src="${item.ImageLink}" alt="drink image">
-    <div class="w-100 d-flex flex-column justify-content-between">
-      <div class="drinks-card-body ms-16">
-        <h4 class="mb-8 mb-md-12">${item.DrinkName}</h4>
-        <ul class="drinks-tag-group mb-8 mb-md-12">
-          ${drinkTagAry[item.id - 1]}
-        </ul>
-        <p class="drinks-card-content mb-24 mb-md-32">${item.Description}</p>
-      </div>
-      <div class="d-flex justify-content-between align-items-end ms-16">
-        <div class="d-flex align-items-center drinkStoreTag">
-          <img src="https://raw.githubusercontent.com/ahmomoz/WhatDrink17/main/assets/images/tri.svg" class="tri" alt="">
-          <p class="bg-primary rounded-2 fw-medium text-white ps-12 pe-10 py-4">${
-            item.StoreName
-          }</p>
-        </div>
-        <a href="#" class="d-block text-primary text-end"><span
-            class="material-symbols-outlined me-2 align-middle">
-            location_on
-          </span>搜尋店家</a>
-      </div>
-  </li>
-`;
+          <li class="drinks-card px-16 py-24 px-md-24">
+          ${btn_tag}
+          <img src="${item.ImageLink}" alt="drink image">
+          <div class="w-100 d-flex flex-column justify-content-between">
+            <div class="drinks-card-body ms-16">
+              <h4 class="mb-8 mb-md-12">${item.DrinkName}</h4>
+              <ul class="drinks-tag-group mb-8 mb-md-12">
+                ${drinkTagAry[item.id - 1]}
+              </ul>
+              <p class="drinks-card-content mb-24 mb-md-32">${item.Description}</p>
+            </div>
+            <div class="d-flex justify-content-between align-items-end ms-16">
+              <div class="d-flex align-items-center drinkStoreTag">
+                <img src="https://raw.githubusercontent.com/ahmomoz/WhatDrink17/main/assets/images/tri.svg" class="tri" alt="">
+                <p class="bg-primary rounded-2 fw-medium text-white ps-12 pe-10 py-4">${
+                  item.StoreName
+                }</p>
+              </div>
+              <a href="#" class="d-block text-primary text-end"><span
+                  class="material-symbols-outlined me-2 align-middle">
+                  location_on
+                </span>搜尋店家</a>
+            </div>
+        </li>
+      `;
     });
     popularDrinkList.innerHTML = str;
   };
@@ -196,7 +180,6 @@ document.addEventListener("DOMContentLoaded", () => {
   // };
 
   //初始化預設飲料卡片函式--------------------------------------------
-
   const drinkRenderData = () => {
     drinkRender();
   };
@@ -204,32 +187,37 @@ document.addEventListener("DOMContentLoaded", () => {
   //收藏愛心CSS樣式函式-----------------------------------------------
   function isCollect() {
     const collectBtn = document.querySelectorAll(".collect-drink-btn"); //抓按鈕class
-    // console.log(collectBtn);//驗證
     //跑 querySelectorAll 陣列
     collectBtn.forEach(function (item) {
       //監聽按鈕
       item.addEventListener("click", function (e) {
-        const drinkId = this.getAttribute("data-drink-id");
-        // console.log(drinkId);
+        if(user_id===null){   //確認是否為會員，不是的話導向會員頁
+          redirectToLogin(); //導向登入頁函數
+        }else{
+          const drinkId = this.getAttribute("data-drink-id");
 
-        //已經收藏時，value 已改成傳送 uncollect，點擊後變為 collected，並移除填滿樣式class，新增外框樣式class
-        if (e.target.value == "collected") {
-          item.value = "uncollect";
-          item.classList.add("fa-regular");
-          item.classList.remove("fa-solid");
-          deleteUserDrinkCollection(user_id, drinkId);
-          Swal.fire("已取消收藏");
-          //還沒收藏時，value 預設傳送 collected，點擊後改傳uncollect，並移除外框樣式class、新增填滿樣式class
-        } else if (e.target.value == "uncollect") {
-          item.value = "collected";
-          item.classList.remove("fa-regular");
-          item.classList.add("fa-solid");
-          addUserDrinkCollection(user_id, drinkId);
-          Swal.fire("收藏成功");
-        }
-        //驗證
-        // console.log("hi");
+          //已經收藏時，value 已改成傳送 uncollect，點擊後變為 collected，並移除填滿樣式class，新增外框樣式class
+          if (e.target.value == "collected") {
+            item.value = "uncollect";
+            item.classList.add("fa-regular");
+            item.classList.remove("fa-solid");
+            deleteUserDrinkCollection(user_id, drinkId);
+            Swal.fire("已取消收藏");
+            //還沒收藏時，value 預設傳送 collected，點擊後改傳uncollect，並移除外框樣式class、新增填滿樣式class
+          } else if (e.target.value == "uncollect") {
+            item.value = "collected";
+            item.classList.remove("fa-regular");
+            item.classList.add("fa-solid");
+            addUserDrinkCollection(user_id, drinkId);
+            Swal.fire("收藏成功");
+          };
+        };
       });
     });
-  }
+  };
 });
+
+//導回登入頁函數-------------------------------------------------------
+function redirectToLogin() {
+  window.location.href = "logIn.html";
+};
